+++ spec/models/organization_spec.rb
@@ -20,6 +20,7 @@ RSpec.describe Organization, type: :model do
   let(:existing_org) { create :organization, organization_name: 'Already Existing Organization' }
 
   it { should have_many :chapters }
+  it { should have_many :students }
 
   describe 'is valid' do
     it 'with a valid, unique name and a unique MLID' do
@@ -121,7 +122,7 @@ RSpec.describe Organization, type: :model do
 
         @chapters = create_list :chapter, 2, organization: @organization_to_delete
         @groups = create_list :group, 2, chapter: @chapters.first
-        @students = create_list :student, 2, group: @groups.first
+        @students = create_list :enrolled_student, 2, groups: [@groups.first], organization: @organization_to_delete
         @lessons = create_list :lesson, 2, group: @groups.first
         @grades = create_list :grade, 2, lesson: @lessons.first
         @deleted_chapter = create :chapter, organization: @organization_to_delete, deleted_at: Time.zone.now
@@ -153,7 +154,7 @@ RSpec.describe Organization, type: :model do
 
         @chapters = create_list :chapter, 2, organization: @organization_to_restore, deleted_at: @organization_to_restore.deleted_at
         @groups = create_list :group, 2, chapter: @chapters.first, deleted_at: @organization_to_restore.deleted_at
-        @students = create_list :student, 2, group: @groups.first, deleted_at: @organization_to_restore.deleted_at
+        @students = create_list :enrolled_student, 2, groups: [@groups.first], deleted_at: @organization_to_restore.deleted_at, organization: @organization_to_restore
         @lessons = create_list :lesson, 2, group: @groups.first, deleted_at: @organization_to_restore.deleted_at
         @grades = create_list :grade, 2, lesson: @lessons.first, deleted_at: @organization_to_restore.deleted_at
         @deleted_chapter = create :chapter, organization: @organization_to_restore, deleted_at: Time.zone.now